Creating/Updating User Review Functionality

+++ CIS560Project/CIS560Project/Models/MovieReview.cs
@@ -12,9 +12,9 @@ namespace CIS560Project.Models
 
         public int ReviewerId { get; }
 
-        public int Score { get; }
+        public double Score { get; }
 
-        public MovieReview(int movieId, int reviewerId, int score)
+        public MovieReview(int movieId, int reviewerId, double score)
         {
             this.MovieId = movieId;
             this.ReviewerId = reviewerId;
